libreria sensores eliminadas, creada libreria general
He eliminado las librerias individuales de sensores para crear una libreria general pudiendo usarla para casi cualquier sensor. Y se ha modificado el programa para que compile todo. Falta arreglar algunos fallos.
