--- identificapersonas_con librerias
@@ -1,18 +1,0 @@
-#include "IdentificarPersonas.h"
-#include <EEPROM.h>
-int numero;
-void setup()
-{
-  Serial.begin(9600);
-  EEPROM.begin(512);
-}
-
-void loop()
-{
-  if (Serial.available() > 1) {
-    numero = Serial.parseInt();
-    leerNumero(numero);
-    mostrarNumero();
-  }
-
-}
