--- libreria_identificarpersonas
@@ -1,16 +1,0 @@
-/*
- * IdentificarPersonas.h - Libreria para calcular Humedad
- */
-
- #ifndef IdentificarPersonas_h
- #define IdentificarPersonas_h
- 
- 
-#include <EEPROM.h>
- #include "Arduino.h"
-
- void inicializarMemoriaEEPROM()
- void leerNumero(int numero);
- void mostrarNumero();
- 
- #endif
